Move Air France parsers under src/parsers/airfrance/
The three parsers (ep5Parser, payParser, nightsAFParser) are AF-specific —
EP5 is an Air France document, the payslip regexes match AF's "BULLETIN DE
PAIE_AIR FRANCE" format, and the nuitées attestation is the AF document.
Grouping them under an airfrance/ namespace makes room for the parallel
src/parsers/transavia/ tree coming next, and makes router.js self-documenting
about which file belongs to which carrier.

  src/parsers/ep5Parser.js       → src/parsers/airfrance/ep5Parser.js
  src/parsers/payParser.js       → src/parsers/airfrance/payParser.js
  src/parsers/nightsAFParser.js  → src/parsers/airfrance/nightsParser.js
                                   (the "AF" suffix is now redundant)

router.js stays at src/parsers/ since it's the airline-agnostic dispatcher.

Mirror the layout under test/: rename test/parsers.test.js → test/parsers/
airfrance.test.js. The three AF integration tests stay grouped in one file
since they share the router-fixture-loadFixture preamble; transavia.test.js
will be its sibling.

Internal import paths updated for the deeper nesting; no behavior change.
